Adding Lombok annotations
To make the classes more transparent and have fewer lines of code, I added Lombok annotations.
