+++ src/main/java/com/mac2work/myfirstproject/webapp/model/City.java
@@ -6,7 +6,15 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity(name = "cities")
 public class City {
 
@@ -17,45 +25,5 @@ public class City {
 	private String name;
 	private Double lat;
 	private Double lon;
-
-	public City() {
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public Double getLat() {
-		return lat;
-	}
-
-	public void setLat(Double lat) {
-		this.lat = lat;
-	}
-
-	public Double getLon() {
-		return lon;
-	}
-
-	public void setLon(Double lon) {
-		this.lon = lon;
-	}
-
-	@Override
-	public String toString() {
-		return "City [id=" + id + ", name=" + name + ", lat=" + lat + ", lon=" + lon + "]";
-	}
 
 }
