+++ src/main/java/com/mac2work/myfirstproject/webapp/model/Plan.java
@@ -4,7 +4,6 @@ import java.time.LocalDate;
 
 import org.hibernate.annotations.GenericGenerator;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -12,7 +11,15 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity(name = "plans")
 public class Plan {
 
@@ -20,95 +27,18 @@ public class Plan {
 	@GeneratedValue(strategy = GenerationType.AUTO, generator = "native")
 	@GenericGenerator(name = "native", strategy = "native")
 	private Long id;
-
 	private LocalDate date;
-
 	private String note;
-
 	private Double successPropability;
-
+	
 	@ManyToOne(fetch = FetchType.LAZY, optional = false)
 	@JoinColumn(name = "city_id", nullable = false)
 	private City city;
-
+	
 	private boolean isDone;
-
+	
 	@ManyToOne(fetch = FetchType.LAZY, optional = false)
 	@JoinColumn(name = "user_id", nullable = false)
 	private User user;
 	
-	public Plan() {}
-	
-	public Plan(LocalDate date, String note, City city, Double successPropability) {
-		this.date = date;
-		this.note = note;
-		this.city = city;
-		this.successPropability = successPropability;
-		this.isDone = false;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public LocalDate getDate() {
-		return date;
-	}
-
-	public void setDate(LocalDate date) {
-		this.date = date;
-	}
-
-	public String getNote() {
-		return note;
-	}
-
-	public void setNote(String note) {
-		this.note = note;
-	}
-
-	public Double getSuccessPropability() {
-		return successPropability;
-	}
-
-	public void setSuccessPropability(Double successPropability) {
-		this.successPropability = successPropability;
-	}
-
-	public City getCity() {
-		return city;
-	}
-
-	public void setCity(City city) {
-		this.city = city;
-	}
-
-	public boolean isDone() {
-		return isDone;
-	}
-
-	public void setDone(boolean isDone) {
-		this.isDone = isDone;
-	}
-	
-	public User getUser() {
-		return user;
-	}
-
-	public void setUser(User user) {
-		this.user = user;
-	}
-
-	@Override
-	public String toString() {
-		return "Plan [id=" + id + ", date=" + date + ", note=" + note + ", successPropability=" + successPropability
-				+ ", city=" + city + ", isDone=" + isDone + ", user=" + user + "]";
-	}
-
-
-
 }
