+++ src/main/java/com/mac2work/myfirstproject/webapp/model/Role.java
@@ -1,0 +1,6 @@
+package com.mac2work.myfirstproject.webapp.model;
+
+public enum Role {
+	USER,
+	ADMIN
+}
