--- src/main/java/com/mac2work/myfirstproject/webapp/model/SecurityUser.java
@@ -1,56 +1,0 @@
-package com.mac2work.myfirstproject.webapp.model;
-
-import java.util.Arrays;
-import java.util.Collection;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public class SecurityUser implements UserDetails{
-
-	private User user;
-	
-	public SecurityUser(User user) {
-		this.user = user;
-	}
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return Arrays.stream(user
-				.getRole()
-				.split(","))
-				.map(SimpleGrantedAuthority::new)
-				.toList();
-	}
-
-	@Override
-	public String getPassword() {
-		return user.getPassword();
-	}
-
-	@Override
-	public String getUsername() {
-		return user.getUsername();
-	}
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
-
-}
